// CRKing7_ADMIN/src/types/banner.type.ts
export interface Banner {
  categoryId: number;
  createdDate: string;
  id: number;
  modifiedDate: string;
  name: string;
  src: string;
  status: number;
}
// const bb ={
//   categoryId
// : 
// 6
// createdDate
// : 
// "10:08:57 24-08-2023"
// id
// : 
// 1
// modifiedDate
// : 
// "10:08:57 24-08-2023"
// name
// : 
// "Banner Home"
// src
// : 
// "slide_home_1.jpg"
// status
// : 
// 1
// }